+++ src/main/java/goorm_runner/backend/post/presentation/post/dto/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package goorm_runner.backend.post.dto;
+package goorm_runner.backend.post.presentation.post.dto;
 
 public record PostUpdateRequest(String title, String content) {
 }
