+++ src/main/java/goorm_runner/backend/post/presentation/post/dto/PostUpdateResponse.java
@@ -1,4 +1,4 @@
-package goorm_runner.backend.post.dto;
+package goorm_runner.backend.post.presentation.post.dto;
 
 import lombok.Builder;
