+++ src/test/java/goorm_runner/backend/comment/presentation/CommentControllerIntegrationTest.java
@@ -11,8 +11,8 @@ import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Comment;
 import goorm_runner.backend.post.domain.model.Post;
-import goorm_runner.backend.post.dto.PostCreateRequest;
 import goorm_runner.backend.post.presentation.comment.dto.CommentCreateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostCreateRequest;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
