+++ src/test/java/goorm_runner/backend/post/application/PostServiceIntegrationTest.java
@@ -7,7 +7,7 @@ import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
-import goorm_runner.backend.post.dto.PostCreateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostCreateRequest;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
