+++ src/test/java/goorm_runner/backend/post/application/PostServiceTest.java
@@ -5,8 +5,8 @@ import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
-import goorm_runner.backend.post.dto.PostCreateRequest;
-import goorm_runner.backend.post.dto.PostUpdateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostCreateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostUpdateRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
