+++ src/test/java/goorm_runner/backend/post/presentation/PostControllerIntegrationTest.java
@@ -9,8 +9,8 @@ import goorm_runner.backend.post.application.post.PostService;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.model.Category;
 import goorm_runner.backend.post.domain.model.Post;
-import goorm_runner.backend.post.dto.PostCreateRequest;
-import goorm_runner.backend.post.dto.PostUpdateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostCreateRequest;
+import goorm_runner.backend.post.presentation.post.dto.PostUpdateRequest;
 import goorm_runner.backend.postlike.application.PostLikeService;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
